requirements.txt
streamlit>=1.22.0
pillow>=9.5.0
numpy>=1.24.3